FadeWorkS: Fix off-by-one error for \ko fades
`koko_da` determines the \ko duration by dividing the "Fade in" value by
the number of word characters and punctuation if the "Fade in" value is
at least 40.

`len` contains the number of word characters and punctuation, but 1 is
erroneously subtracted from it before the division, resulting in
durations that are too long to fit the "Fade in" duration. The fix is
to remove the subtraction and simply divide by `len`.